Refactor parseRequestParameters
