+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/SignatureRequestParameters.kt
@@ -1,4 +1,4 @@
-package at.asitplus.openid.rqes
+package at.asitplus.openid
 
 import at.asitplus.dif.rqes.collection_entries.DocumentDigestEntries.CscDocumentDigest
 import at.asitplus.dif.rqes.collection_entries.DocumentDigestEntries.OAuthDocumentDigest
@@ -7,8 +7,6 @@ import at.asitplus.dif.rqes.enums.ConformanceLevelEnum
 import at.asitplus.dif.rqes.enums.SignatureFormat
 import at.asitplus.dif.rqes.enums.SignatureQualifierEnum
 import at.asitplus.dif.rqes.enums.SignedEnvelopeProperty
-import at.asitplus.openid.AuthorizationDetails
-import at.asitplus.openid.OpenIdConstants
 import at.asitplus.signum.indispensable.Digest
 import at.asitplus.signum.indispensable.X509SignatureAlgorithm
 import at.asitplus.signum.indispensable.asn1.Asn1Element
@@ -69,7 +67,7 @@ data class SignatureRequestParameters(
 
     @SerialName("clientData")
     val clientData: String?,
-) {
+) : RequestParameters {
     fun toAuthorizationDetails(): AuthorizationDetails =
         AuthorizationDetails.CSCCredential(
             credentialID = this.clientId,
