+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/AuthenticationRequest.kt
@@ -5,7 +5,6 @@ import at.asitplus.catching
 import at.asitplus.dif.rqes.serializers.UrlSerializer
 import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.openid.JwsSignedSerializer
-import at.asitplus.openid.rqes.SignatureRequestParameters
 import io.ktor.http.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
