+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/SignatureRequestParameterFrom.kt
@@ -3,7 +3,7 @@ package at.asitplus.wallet.lib.oidc
 import at.asitplus.catching
 import at.asitplus.dif.rqes.Serializer.UrlSerializer
 import at.asitplus.openid.JwsSignedSerializer
-import at.asitplus.openid.rqes.SignatureRequestParameters
+import at.asitplus.openid.SignatureRequestParameters
 import io.ktor.http.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
