+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/RqesWalletService.kt
@@ -6,7 +6,7 @@ import at.asitplus.dif.rqes.RqesConstants
 import at.asitplus.dif.rqes.SignDocParameters
 import at.asitplus.dif.rqes.SignHashParameters
 import at.asitplus.openid.AuthenticationRequestParameters
-import at.asitplus.openid.rqes.SignatureRequestParameters
+import at.asitplus.openid.SignatureRequestParameters
 import at.asitplus.signum.indispensable.X509SignatureAlgorithm
 import at.asitplus.wallet.lib.oauth2.OAuth2Client
 import com.benasher44.uuid.uuid4
@@ -18,7 +18,7 @@ class RqesWalletService(
 ) {
 
     suspend fun createOAuth2AuthenticationRequest(
-        rqesRequest: at.asitplus.openid.rqes.SignatureRequestParameters,
+        rqesRequest: SignatureRequestParameters,
         credentialId: ByteArray,
     ): AuthenticationRequestParameters =
         oauth2Client.createAuthRequest(
@@ -45,7 +45,7 @@ class RqesWalletService(
         )
 
     suspend fun createSignHashRequestParameters(
-        rqesRequest: at.asitplus.openid.rqes.SignatureRequestParameters,
+        rqesRequest: SignatureRequestParameters,
         credentialId: String,
         sad: String,
     ): CSCSignatureRequestParameters = SignHashParameters(
